Add in support for gateway SSH connections
This is sometimes known as multi-hop ssh connections and is achieved using the pseudo tty ssh feature.  The end result is a single command that creates an SSH connection through one or more hosts into the destination host.  This is especially beneficial for connecting through multiple network layers (VLAN's or VPC's) 

Usage:
  -g BASTIONHOST
  -g USERNAME@BASTIONHOST
  -g "USERNAME@BASTIONHOST -i PATH/TO/KEY/FOR/BASTION/CONNECTION"