bugfix(#79): Only split extra param into 2 pieces
Many options to the "ssh -o" flag come in key-value pairs
separated by "=" signs. Currently, attempting to pass such
pairs through the i2cssh "-X" flag will fail, because we
are splitting up the entire argument based on the "=" sign.

E.g. "i2cssh -Xo=StrictHostKeyChecking=no" results in:
     ssh -o StrictHostKeyChecking no
instead of:
     ssh -o StrictHostKeyChecking=no

By taking advantage of the optional second argument to the
Ruby string #split method, we can ensure we only split on
the first occurrence of the "=" sign, which is actually
what we wanted all along.