Version 0.0.9
Forgot to implement the automatic full screen when cols >= 12 and rows
>= 16. This was hard coded in the ruby version and the logic was there
but we never acted on it. This is still a hack, since the amount of rows
and cols is based on the size of the newly spawned window, which can
differ based on the profile used (default is 80x25). The proper thing to
do is to figure out how many characters a terminal needs to be before
iTerm2 disallows another split, check the sise of the new window and
then act accordingly (either resize or go full screen). But we'll save
that for later.