Version 0.0.14: add --custom-command
This release adds a `-x or --custom-command` option as reqested in #118.
This allows for running a different command than `ssh`, e.g.
`kubectl "kubectl exec -it {host} -- /bin/bash"` where `{host}` is
substituted by the actual host from the config.

This also works from the config file on either global or cluster level