Version 0.0.19: adds --exec/-E option to close pane after ssh exits. Fixes #127
This commit adds --exec and -E options to exec ssh in the shell (and
thus replacing the shell), rather than just running the command. This
causes the pane to close after exiting ssh, rather than returning to the
shell. Note that this causes any output to disappear. This option is
disabled by default.

Additionally this solves a bug that caused config params not to be
handled the same as command line parameters.