Version 0.1.0: adds test (and big refactor)
This commit finally adds tests! Because of this, I've also restructured
the project a bit and split some code up into bin/ and a proper lib.

And lastly, I refactored the way that options are applied to each other,
since the test actually exposed bugs and missing options in config.