Version 0.2.0
This changes the behavior of what shell to use for new sessions slightly. The default behavior
used to be that /bin/bash (or rather /usr/bin/env bash) was started, but since MacOS changed
its default shell to zsh a while back, this causes issues for people using zsh as their default
shell. Changing the default shell for i2cssh to zsh would break people that still use bash,
so the introduced behavior is that the current shell that the user uses will be used, unless
otherwise specified in configuration (through `-S` or `shell:` in yaml config).

Note that `-S` takes the name of a shell, not the full path.